Add files via upload
This notebook shows the fully checked and tested nuances of the system. There are positive sides to the proposed task-oriented conversation management system, but there are also limitations. So what we are highlighting is that there is a possibility of building a task-oriented conversation system that can utilize the emotional state of the user as perceived through the semantic meaning of user responses and the user engagement that is also perceived through the speech modality (the text-based response of the user). Through this research, we proved it is worth considering the incorporation of "emotional cues of the user" and "engagement of the user" for a task-oriented conversation management system to provide a good user experience in the text-based interaction scenario (for Human-robot interaction-based applications, it can be adopted as a voice based interactive system)